% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{convertir_dates}
\alias{convertir_dates}
\title{Converteix format TEXT A DATA}
\usage{
convertir_dates(
  d = "dt_plana",
  taulavariables = "conductor",
  campdata = "dates"
)
}
\arguments{
\item{d}{Base de dades}

\item{taulavariables}{Variables que volem canviar}

\item{campdata}{dates}
}
\value{
Data
}
\description{
Converteix format TEXT A DATA ("19722207"->"1972-02-22")
}
\examples{
camp=c("idp","dtindex","sexe","dnaix","situacio","entrada",
      "sortida", "INCLUSIO.DM2","DG.HTA","DG.IC",
      "cHDL.valor","cLDL.valor","cT.valor","GLICADA.valor","IMC.valor","data2")
factor=c("","","","","","","",1,1,1,"","","","","",1)
dates=c("","","","","","","","","","","","","","","",1)
conductor<-data.frame(camp,factor,dates)
dt_plana_test$data2<-"19722207"
K<-convertir_dates(d=dt_plana_test,taulavariables=conductor,campdata="dates")
K
}
