% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{data.random}
\alias{data.random}
\title{Data Random entre dues dates}
\usage{
data.random(dataini = 20120101, datafi = 20121231)
}
\arguments{
\item{dataini}{Data inicial}

\item{datafi}{Data final}
}
\value{
Una data random tiupus "2012-10-23 20:57:28 CEST"
}
\description{
Data Random entre dues dates (dataini = 20120101, datafi = 20121231)
}
\examples{
data<-data.random(20120101,20121231)
data
}
