% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{data_convert_numeric}
\alias{data_convert_numeric}
\title{Funcio que converteix de una data numeric (15784) a Date}
\usage{
data_convert_numeric(x)
}
\arguments{
\item{x}{Data numeric}
}
\value{
Data "2013-03-20"
}
\description{
Funcio que converteix de numeric (15784) a Date "2013-03-20"
}
\examples{
k<-data_convert_numeric(15784)
k
}
