% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{data_convert_text}
\alias{data_convert_text}
\title{Converteix data caracter () a una data ymd}
\usage{
data_convert_text(x)
}
\arguments{
\item{x}{Data caracter}
}
\value{
Data ymd
}
\description{
Converteix data caracter ("37712") a date ymd () "2003-04-01"
}
\examples{
k<-data_convert_text("37712")
k
}
