% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{data_convert_UTC}
\alias{data_convert_UTC}
\title{Funcio que converteix UTC data a date ymd}
\usage{
data_convert_UTC(x)
}
\arguments{
\item{x}{UTC data}
}
\value{
Date ymd
}
\description{
Funcio que converteix UTC data a date ymd
}
\examples{
k<-data_convert_UTC("1970-01-01")
k
}
