% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{dataSPSS_to_Rdata}
\alias{dataSPSS_to_Rdata}
\title{Passa data de SPSS a Rdata}
\usage{
dataSPSS_to_Rdata(x)
}
\arguments{
\item{x}{Data de SPSS}
}
\value{
Data Rdata
}
\description{
Passa data de SPSS a Rdata
}
\examples{
k<-dataSPSS_to_Rdata(666)
k
}
