% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatGe.R
\name{dataUTC_to_Rdata}
\alias{dataUTC_to_Rdata}
\title{Data R Lubridate a partir de data UTC}
\usage{
dataUTC_to_Rdata(x, dt)
}
\arguments{
\item{x}{Variablesde data}

\item{dt}{Base de dades}
}
\value{
Base de dades amb Data R Lubridate
}
\description{
Data R Lubridate a partir de data UTC
}
\examples{

dt_plana
x=c("dtindex","dnaix","entrada","sortida", "INCLUSIO.DM2")
dataUTC_to_Rdata(x,dt_plana)
}
