% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetFormatGe.R
\name{etiquetar}
\alias{etiquetar}
\title{Etiquetar els noms de les variables}
\usage{
etiquetar(
  d = "dadestotal",
  taulavariables = "variables_R.xls",
  camp_descripcio = "descripcio",
  ...
)
}
\arguments{
\item{d}{Base de dades}

\item{taulavariables}{conductor}

\item{camp_descripcio}{camp del conductor}

\item{...}{altres funcions}
}
\value{
Etiquetar
}
\description{
Etiquetar les variables de les dades a partir d'un conductor.
}
\examples{
dades_etiquet<-etiquetar(dt_plana_test,conductor1_test, "descripcio")
Hmisc::label(dt_plana_test)
Hmisc::label(dades_etiquet)
}
