% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetFormatGe.R
\name{etiquetar_valors}
\alias{etiquetar_valors}
\title{Etiquetar valors de les variables}
\usage{
etiquetar_valors(
  dt = "dades",
  variables_factors = "conductor_variables",
  fulla = "etiquetes",
  camp_etiqueta = "etiqueta",
  missings = F,
  new_vars = F,
  sufix = ".2"
)
}
\arguments{
\item{dt}{Base de dades}

\item{variables_factors}{Conductor}

\item{fulla}{Si tenim el conductor a una pestanya}

\item{camp_etiqueta}{Camp de les etiquetes}

\item{missings}{Missing}

\item{new_vars}{Nova variable}

\item{sufix}{Sufix}
}
\value{
Dataframe dades,  conductor_variables
}
\description{
Etiquetar valors de les variables a partir d'un conductor -pestanya etiquetes-
}
\examples{
camp=c("sexe","sexe","situacio","situacio","situacio")
valor=c("H","D","A","D","T")
etiqueta1=c("1.Dona","2.Home","1.Actiu","2.Difunt","3.Trasllat")
etiqueta2=c("1.Mujer","2.Hombre","1.Activo","2.Difunto","3.Traslado")
conductor2<-data.frame(camp,valor,etiqueta1,etiqueta2)


table(dt_plana$sexe) 

dades_valors_etiquet<-etiquetar_valors(
dt=dt_plana,
variables_factors=conductor2,
fulla="etiquetes",
camp_etiqueta="etiqueta1",
missings=FALSE, 
new_vars=FALSE,
sufix=".2")

table(dades_valors_etiquet$sexe) 

}
