% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetFormatGe.R
\name{etiquetar_valors_patro}
\alias{etiquetar_valors_patro}
\title{Etiquetar valors de camps dun data frame}
\usage{
etiquetar_valors_patro(
  dt,
  taulavariables,
  text_opcions = "select_choices_or_calculations",
  sep = "|",
  ...
)
}
\arguments{
\item{dt}{Data frame}

\item{taulavariables}{Conductor}

\item{text_opcions}{de la columna referent al text on hi ha valors i etiquetes}

\item{sep}{Separador ("|")}

\item{...}{altres funcions exemple fulla del conductor}
}
\value{
tibble
}
\description{
Donat conductor amb patró (valor i etiqueta:  Ex "valor1, etiqueta1 | valor2, etiqueta 2) retorna tibble camps recodificats (valor, etiqueta) corresponent a info separada per un caracter sep ("|")
}
