% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoritzarFormatGe.R
\name{factoritzar.NO.SI}
\alias{factoritzar.NO.SI}
\title{Factoritzar (No/Si)}
\usage{
factoritzar.NO.SI(
  dt = "dadesDF",
  columna = "factor",
  taulavariables = "variables_FELIPE.xls",
  ...
)
}
\arguments{
\item{dt}{Base de dades}

\item{columna}{La columna del conductor a on farem la factoritzacio}

\item{taulavariables}{Conductor}

\item{...}{Altres funcions}
}
\value{
NO.Si
}
\description{
A tots aquells valors dummis, el "0"-->"No" i "1"-->"Si"
}
\examples{

dt_plana2<-dplyr::mutate_at(dt_plana, dplyr::vars(  dplyr::starts_with("DG.") ),  
dplyr::funs( dplyr::if_else(.==0  | is.na(.)  ,0,1)))
dt_plana2<-dplyr::mutate_at(dt_plana2, dplyr::vars(  dplyr::starts_with("INCLUSIO.") ), 
dplyr::funs( dplyr::if_else(.==0  | is.na(.)  ,0,1)))
dt_plana2
dt_plana2<-factoritzar.NO.SI(dt_plana2,"factor",conductor1)
dt_plana2

}
