% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoritzarFormatGe.R
\name{factoritzar}
\alias{factoritzar}
\title{Factoritzar FormatGe}
\usage{
factoritzar(dt = "dades", variables = c("grup", "situacio"))
}
\arguments{
\item{dt}{Base de dades plana}

\item{variables}{Vector  a on posem les variables a factoritzar}
}
\value{
La base de dades amb les variables factoritzades
}
\description{
Factoritza una llista de variables donades a la teva base de dades
}
\examples{

class(dt_plana_test$situacio)
dt_plana2<-factoritzar(dt_plana_test,variables=c("grup","situacio"))
class(dt_plana2$situacio)

}
