% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGe.R
\docType{package}
\name{formatGe}
\alias{formatGe}
\title{formatGe Package}
\description{
FormatGe contain functions to format and generate variables to data table through structural external file (Conductor)
}
\author{
Jordi Real \email{jordireal@gmail.com}
}
