% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFormatGe.R
\name{formula_table1}
\alias{formula_table1}
\title{Formula Table.}
\usage{
formula_table1(
  x = "taula1",
  y = "",
  eliminar = c("IDP"),
  a = "",
  taulavariables = "variables.xls",
  dt = NA,
  ...
)
}
\arguments{
\item{x}{Indicador de cadena del camp del fitxer extern amb indicador de la posició de la fórmula (1,2....)}

\item{y}{Indicador de resposta de cadena (default "."): ". ~ x1 +x2 + x3 "}

\item{eliminar}{Caracter que indica si alguna variable ha de ser eliminada}

\item{a}{Caracter , indica si cal afegir alguna variable en la primera posició}

\item{taulavariables}{Conductor , a on triem l'orde de les variables de la formula}

\item{dt}{Data.frame o base de dades}

\item{...}{Altres funcions}
}
\value{
Formula_table1
}
\description{
Retorna formula per table1::table1 segons llista de varibles
}
\examples{
Hmisc::label(iris)
conductor_iris<-data.frame(camp=names(iris),formu=c(1,2,3,4,5))
formula_iris2<-formula_table1("formu",taulavariables = conductor_iris)
formula_iris2
}
