% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulaFormatGe.R
\name{formula_text2}
\alias{formula_text2}
\title{Formula FormatGe.}
\usage{
formula_text2(
  x = "taula1",
  y = "resposta",
  eliminar = c("IDP"),
  a = "",
  taulavariables,
  dt = NA
)
}
\arguments{
\item{x}{String indicator of field from external file with indicator the position of of formula (1,2....).}

\item{y}{string indicator of response (default "."): ". ~ x1 +x2 + x3 "}

\item{eliminar}{character that indicate if some variable has to be clean}

\item{a}{character that indicate if some variable has to be add in the first position}

\item{taulavariables}{excel file with number indicator field (discrete number 1:Inf)}

\item{dt}{data.frame}
}
\value{
Data.frame or tibble with labeled variables
}
\description{
Retorna Data.frame or tibble with labeled variables
}
\examples{

Hmisc::label(iris)
conductor_iris<-data.frame(camp=names(iris),formu=c(1,2,3,4,5))
formula_iris<-formula_text("formu",taulavariables = conductor_iris)
formula_iris
}
