% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetFormatGe.R
\name{generar_dt_conductor_labels}
\alias{generar_dt_conductor_labels}
\title{Generar conductor d etiquetes}
\usage{
generar_dt_conductor_labels(
  chart_camp = "rb_grupo",
  path_conductor,
  text_opcions = "select_choices_or_calculations",
  sep = "|",
  ...
)
}
\arguments{
\item{chart_camp}{Camp}

\item{path_conductor}{Conductor}

\item{text_opcions}{de la columna referent al text on hi ha valors i etiquetes}

\item{sep}{Separador ("|")}

\item{...}{altres funcions exemple fulla del conductor}
}
\value{
tibble
}
\description{
Retorna tibble amb dos camps (valor, etiqueta) corresponent a info separada per un caracter sep ("|")
}
