% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodiFormatGe.R
\name{generar_intervals}
\alias{generar_intervals}
\title{Generar intervals amb talls a variables continues}
\usage{
generar_intervals(
  dt = "dades",
  vars = "ajust4",
  taulavariables = "conductor1",
  missing = "Unkown",
  g = 3
)
}
\arguments{
\item{dt}{Base de dades}

\item{vars}{Variables}

\item{taulavariables}{Conductor}

\item{missing}{Missing}

\item{g}{Numero de talls}
}
\value{
Retorna intervals
}
\description{
Generar intervals a partir d'un conductor i un tall definit
}
\examples{

kk<-generar_intervals(dt=dt_plana_test,
vars="ajust4",
taulavariables=conductor1_test,
missing="Unkown",
g=2)

kk
}
