% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelData.R
\name{labelData}
\alias{labelData}
\title{Label variables according external file
Open a excel file and apply label to variables frome data base}
\usage{
labelData(data, taulavariables, camp_descripcio)
}
\arguments{
\item{data}{data.frame}

\item{taulavariables}{excel file with description field}

\item{camp_descripcio}{Field with label in excel file}
}
\value{
Data.frame or tibble with labeled variables
}
\description{
Label variables according external file
Open a excel file and apply label to variables frome data base
}
