% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTable.R
\name{labelTable}
\alias{labelTable}
\title{Exchange label for name of variable according external file
Open a excel file and apply label to variables from table}
\usage{
labelTable(data, taulavariables, camp_descripcio, camp)
}
\arguments{
\item{data}{data.frame}

\item{taulavariables}{excel file with description field}

\item{camp_descripcio}{Field with label in excel file}

\item{camp}{character name field}
}
\value{
Data.frame or tibble with labeled variables
}
\description{
Exchange label for name of variable according external file
Open a excel file and apply label to variables from table
}
