% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelData.R
\name{levelData}
\alias{levelData}
\title{Apply levels to values in variables according external file}
\usage{
levelData(data, taulavariables, fulla, camp_etiqueta)
}
\arguments{
\item{data}{data.frame}

\item{taulavariables}{excel file with description field}

\item{fulla}{sheet in excel file with description field}

\item{camp_etiqueta}{field with value and label in excel file}
}
\value{
Data.frame or tibble with leveled variables
}
\description{
Apply levels to values in variables according external file
}
