% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetFormatGe.R
\name{llistaNomenada}
\alias{llistaNomenada}
\title{Llista Nomenada}
\usage{
llistaNomenada(...)
}
\arguments{
\item{...}{Altres funcions}
}
\value{
Llista Nomenada
}
\description{
Retorna llista nomenada amb els mateixos noms dels objectes que inclou
}
