% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodiFormatGe.R
\name{missings_to_level}
\alias{missings_to_level}
\title{Missings_to_level}
\usage{
missings_to_level(dades, variable = "popes")
}
\arguments{
\item{dades}{Base de dades}

\item{variable}{Variables}
}
\value{
Retorna missings
}
\description{
Recodificar  missings_to_level
}
\examples{
k1<-missings_to_level(dt_plana,variable="DG.HTA")
k1
}
