% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netejarFormatGe.R
\name{netejar.accents.variables}
\alias{netejar.accents.variables}
\title{Elimina accents dels noms de les variables}
\usage{
netejar.accents.variables(dt = "LIPOS_EORTEGA")
}
\arguments{
\item{dt}{Base de dades}
}
\value{
Base de dades  sense accents
}
\description{
Elimina accents dels noms de les variables
}
\examples{
dt_plana$"sexé2"<-1
kk<-netejar.accents.variables(dt_plana)
stats::variable.names(dt_plana)
stats::variable.names(kk)
}
