% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netejarFormatGe.R
\name{netejar.noms.variables}
\alias{netejar.noms.variables}
\title{Netejar els noms de les variables}
\usage{
netejar.noms.variables(dt = "hoolywod")
}
\arguments{
\item{dt}{Base de dades}
}
\value{
Base de dades amb la neteja de noms
}
\description{
Netejar els noms de les variables ["/","(",")","/"],-> ""  [" ","-"]-> "_" ,..
}
\examples{
dt_plana$"var//"<-1
kk<-netejar.noms.variables(dt_plana)
stats::variable.names(dt_plana)
stats::variable.names(kk)
}
