% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netejarFormatGe.R
\name{netejar_espais}
\alias{netejar_espais}
\title{Elimina espais dels noms de les variables}
\usage{
netejar_espais(dt = "dades")
}
\arguments{
\item{dt}{Base de dades}
}
\value{
Base de dades elimina espais
}
\description{
Elimina espais dels noms de les variables
}
\examples{
dt_plana$"  var3  "<-69
kk<-netejar_espais(dt=dt_plana)
stats::variable.names(dt_plana)
stats::variable.names(kk)
}
