% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodiFormatGe.R
\name{recode_to_missings}
\alias{recode_to_missings}
\title{Recodificar missings}
\usage{
recode_to_missings(
  dt = "dades",
  taulavariables = "conductor_variables",
  rang = "rang_valid",
  data_long = FALSE,
  ...
)
}
\arguments{
\item{dt}{Base de dades}

\item{taulavariables}{Conductor}

\item{rang}{Rang}

\item{data_long}{Data Long}

\item{...}{Altres funcions}
}
\value{
Retorna dades amb la recodificacio de missings
}
\description{
Recodificar rangs de valors que cauen fora de interval es transforman en missings
}
\examples{
conductor1_test
dt_plana_test
kk<-recode_to_missings(dt=dt_plana_test,taulavariables=conductor1_test,rang="rang_valid")
kk
}
