% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodiFormatGe.R
\name{recodificar}
\alias{recodificar}
\title{Recodificar les variables que escollim}
\usage{
recodificar(
  dt = "dades",
  taulavariables = "VARIABLES.xls",
  criteris = "recode1",
  missings = F,
  prefix = NA,
  ...
)
}
\arguments{
\item{dt}{Base de dades}

\item{taulavariables}{Conductor}

\item{criteris}{Criteris ex:: 1/3/5}

\item{missings}{Missing}

\item{prefix}{Posar un prefix a la nova variable}

\item{...}{Altres funcions}
}
\value{
Retorna les dades amb la recodificacio
}
\description{
Recodificar les variabñes que escollim a partir del parametre "criteris"
}
\examples{
k<-recodificar(dt=dt_plana_test, 
              taulavariables=conductor1_test,
              criteris="recode",
              missings=FALSE,
              prefix="_recode_")
              
k
}
