% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcat.R
\name{refcat}
\alias{refcat}
\title{Canviar i definir categoria de referencia}
\usage{
refcat(DF = "iris", conductor = "conductor_iris", ref = "ref_cat", ...)
}
\arguments{
\item{DF}{Base de dades}

\item{conductor}{conductor}

\item{ref}{categoria de referencia}

\item{...}{altres parametres}
}
\value{
Etiquetar
}
\description{
Canviar i definir categoria de referencia en un llistat de variables posades en un conductor.
}
