% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taula_plana_test2.R
\docType{data}
\name{Taula_plana_test2}
\alias{Taula_plana_test2}
\title{Bases de dades de Poblacio}
\format{
Una data frame amb 30 files i 165 variables
\describe{
\item{idp}{idp (1--30)}
\item{dtindex}{dtindex ()}
\item{DG.ALCOHOL}{DG.ALCOHOL ()}
\item{DG.ARTPER}{DG.ARTPER ()}
\item{DG.AVC}{DG.AVC ()}
\item{DG.DM_ALTRES}{DG.DM_ALTRES ()}
}
}
\usage{
Taula_plana_test2
}
\description{
Bases de dades de Poblacio
}
\keyword{datasets}
