% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatGe.R
\docType{package}
\name{FormatGe}
\alias{FormatGe}
\title{FormatGe Package}
\description{
FormatGe conté funcions per formatar i generar variables a la taula de dades mitjancant un sitema de CONDUCTORS.
}
\author{
Jordi Real, Ray Puig \email{-jordireal@gmail.com-}
}
