% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatGe.R
\docType{package}
\name{FormatGe}
\alias{FormatGe}
\title{FormatGe Package}
\description{
FormatGe contain functions to format and generate variables to data table through structural external file (Conductor)
}
\author{
Jordi Real, Ray Puig \email{jordireal@gmail.com}
}
