% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTable.R
\name{labelTable}
\alias{labelTable}
\title{Exchange label for name of variable according external file
Open a excel file and apply label to variables from table}
\usage{
labelTable(
  data = resumtotal,
  taulavariables = "variables_R.xls",
  camp_descripcio = "descripcio",
  camp = "variable"
)
}
\arguments{
\item{data}{data.frame}

\item{taulavariables}{excel file with description field}

\item{camp_descripcio}{Field with label in excel file}
}
\value{
Data.frame or tibble with labeled variables
}
\description{
Exchange label for name of variable according external file
Open a excel file and apply label to variables from table
}
